poprzednia wersja/WH_Slot_booking_python_API/backend/requirements.txt

fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib
